Split testbed window tests into 2 modules.
